Dix the checkdoc warnings.
Fixed the warnings while trying to keep the same description.

It's not always easy to prevent these warnings while keeping the exact
phrasing, so in some of them I re-organized the sentences.  Those are
suggestions.

Also I found that using \`C-x'  (with the single backslash) silences the
warning and does the proper rendering (under Emacs 30). That seems to comply
with what the manual states for \`KEYSEQ'.